obs-ffmpeg: Move muxer structure/funcs to header
(Jim) Allows the ability to use it for more output types.
